Update README.md
This script is for data cleaning and analysis specifically designed for grocery store datasets. It includes the conversion of timestamps to datetime objects, normalization of coordinate data, and mapping of grocery items to defined areas within a store. The tool filters data by specific date ranges, assigns unique path IDs to customer movements, and exports cleaned data for further analysis. It utilizes pandas for data manipulation, geopandas for geospatial analysis, and matplotlib for data visualization, making it a comprehensive solution for grocery store data management.